requirements.txt
 openai
 ollama
 anthropic                                                                      
 termcolor                                                                   
 prompt_toolkit                                                              
 rich
 python-dotenv
 colorlog
